SVN r2980: Added logic for computing satellite positions as a function of time. This required adding quadratic regression capabilities to linearalgebra
Added Novatel test utility (work in progress), which led me to create a PC/Qt version of Uart.h. This led to lots of problems related to the definition of PARITY. Apparently windows uses those same enumerations. Renamed all instances of PARITY_... to UART_PARITY_... to deal with that. Think I got it everywhere.