SVN r3382: Removed the separate enumeration for board types in the debug packet. Board source is now given using the same enumeration as other packets. This breaks interpretation of the debug packet, but since no user is using that it shouldn't be a problem.
Large re-organization of the Orion Public Protocol. No actual changes in the packet designs, but the xml has been changed to take advantage of the 1.5.0 version of Protogen which has nicer documentation capabilities. I verified that the 1.4.5.b protogen (current in the repo) has no complaints with the updated xml.